UpdateBank

+++ src/main/java/com/example/FoodTourApp/DTO/WalletDTO/WalletTransactionResponseDTO.java
@@ -1,10 +1,12 @@
 package com.example.FoodTourApp.DTO.WalletDTO;
 
+import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -17,9 +19,13 @@ public class WalletTransactionResponseDTO {
     private Integer userId;
     private String userFullName;
     private String transactionType;
-    private Double amount;
-    private Double balanceBefore;
-    private Double balanceAfter;
+
+    private BigDecimal amount;
+
+    private BigDecimal balanceBefore;
+
+    private BigDecimal balanceAfter;
+
     private Integer orderId;
     private String orderNumber;
     private String description;
